--- a/rust_tonic_mt_bench/.cargo/config
+++ b/rust_tonic_mt_bench/.cargo/config
@@ -1,0 +1,2 @@
+[build]
+rustflags = [ "-C", "target-cpu=native" ]