--- a/rust_tonic_mt_bench/src/main.rs
+++ b/rust_tonic_mt_bench/src/main.rs
@@ -23,8 +23,24 @@ impl Greeter for MyGreeter {
     }
 }
 
-#[tokio::main]
-async fn main() -> Result<(), Box<dyn std::error::Error>> {
+fn main() -> Result<(), Box<dyn std::error::Error>> {
+    let cpus = std::env::var("GRPC_SERVER_CPUS")
+        .map(|v| v.parse().unwrap())
+        .unwrap_or(1);
+    
+    println!("Running with {} threads", cpus);
+
+    // Esentially the same as tokio::main, but with number of threads set to 
+    // avoid thrashing when cggroup limits are applied by Docker.
+    tokio::runtime::Builder::new_multi_thread()
+        .worker_threads(cpus)
+        .enable_all()
+        .build()
+        .unwrap()
+        .block_on(serve())
+}
+
+async fn serve() -> Result<(), Box<dyn std::error::Error>> {
     let addr = "0.0.0.0:50051".parse().unwrap();
     let greeter = MyGreeter::default();