@mike-barber 2021-04-25
It's a pleasure, and thanks for merging it, @LesnyRumcajs 
It'll be interesting to see what difference it makes on your hardware. If it's not doing what it's supposed to, we can look at it again :)